Update developers-guide-v4.md
I have added notes for the load-balanced setups and explaining some problems we can face.

I have only included this in the v13 docs because is the one that I have tested, but probably it will also apply for the rest. Let me know if you want me to add it to the rest.